Add Note content helper
