+++ Simplenote/Classes/Note+Properties.swift
@@ -39,62 +39,3 @@ extension Note {
         }
     }
 }
-
-// MARK: - Content
-//
-extension Note {
-
-    /// Range of title in content
-    ///
-    var titleRange: NSRange {
-        guard let content = content else {
-            return NSRange(location: NSNotFound, length: 0)
-        }
-
-        let nsContent = content as NSString
-        let fullRange = nsContent.fullRange
-
-        let firstCharacterRange = nsContent.rangeOfCharacter(from: CharacterSet.whitespacesAndNewlines.inverted,
-                                                           options: [],
-                                                           range: fullRange)
-        guard firstCharacterRange.location != NSNotFound else {
-            return firstCharacterRange
-        }
-
-        let newlineSearchRange = NSRange(location: firstCharacterRange.location,
-                                         length: fullRange.length - firstCharacterRange.location)
-        let newlineRange = nsContent.rangeOfCharacter(from: .newlines, options: [], range: newlineSearchRange)
-
-        guard newlineRange.location != NSNotFound else {
-            return newlineSearchRange
-        }
-
-        return NSRange(location: firstCharacterRange.location,
-                       length: newlineRange.location - firstCharacterRange.location)
-    }
-
-    /// Range of body in content
-    ///
-    var bodyRange: NSRange {
-        let titleRange = self.titleRange
-        guard titleRange.location != NSNotFound, let content = content else {
-            return titleRange
-        }
-
-        let nsContent = content as NSString
-        let fullRange = nsContent.fullRange
-
-        let untrimmedBodyRange = NSRange(location: NSMaxRange(titleRange),
-                                         length: fullRange.length - NSMaxRange(titleRange))
-
-        let firstCharacterRange = nsContent.rangeOfCharacter(from: CharacterSet.whitespacesAndNewlines.inverted,
-                                                             options: [],
-                                                             range: untrimmedBodyRange)
-        guard firstCharacterRange.location != NSNotFound else {
-            return firstCharacterRange
-        }
-
-        return NSRange(location: firstCharacterRange.location,
-                       length: content.fullRange.length - firstCharacterRange.location)
-    }
-}
