+++ SimplenoteTests/NoteContentHelperTests.swift
@@ -1,0 +1,9 @@
+import XCTest
+@testable import Simplenote
+
+
+// MARK: - NoteContentHelperTests Tests
+//
+class NoteContentHelperTests: XCTestCase {
+    
+}
